Update vim plugins

+++ vim/bundle/LanguageClient-neovim/tests/test.sh
@@ -5,11 +5,18 @@ set -o xtrace
 dir=$(dirname $(dirname $(realpath $0)))
 cd $dir
 
-LOG="${TMP:-/tmp}"/LanguageClient.log
-LOG_SERVER="${TMP:-/tmp}"/LanguageServer.log
+LOG=~/.local/share/nvim/LanguageClient.log
 
+curl -fLo tests/data/.vim/autoload/plug.vim --create-dirs \
+    https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
+
+nvim -n -u tests/data/vimrc --headless +PlugInstall +qa
 rm -f /tmp/nvim-LanguageClient-IntegrationTest
-NVIM_LISTEN_ADDRESS=/tmp/nvim-LanguageClient-IntegrationTest nvim -n -u tests/data/vimrc --headless &
+if [[ "$TMUX" ]]; then
+    tmux split-window 'NVIM_LISTEN_ADDRESS=/tmp/nvim-LanguageClient-IntegrationTest nvim -n -u tests/data/vimrc'
+else
+    NVIM_LISTEN_ADDRESS=/tmp/nvim-LanguageClient-IntegrationTest nvim -n -u tests/data/vimrc --headless &
+fi
 PID=$!
 sleep 1s
 
@@ -18,8 +25,9 @@ ret=$?
 
 if [[ $ret != 0 ]]; then
     cat $LOG
-    cat $LOG_SERVER
 fi
 
-kill $PID
+if [[ -n $PID ]]; then
+    kill $PID
+fi
 exit $ret
